Add large EMC values and update monitor

+++ src/main/java/net/exohayvan/dissolver_enhanced/common/values/EmcValueOverride.java
@@ -1,27 +1,28 @@
 package net.exohayvan.dissolver_enhanced.common.values;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 public final class EmcValueOverride {
-    private final Map<String, Integer> items;
-    private final Map<String, Integer> tags;
+    private final Map<String, BigInteger> items;
+    private final Map<String, BigInteger> tags;
 
-    public EmcValueOverride(Map<String, Integer> items, Map<String, Integer> tags) {
+    public EmcValueOverride(Map<String, BigInteger> items, Map<String, BigInteger> tags) {
         this.items = immutableCopy(items);
         this.tags = immutableCopy(tags);
     }
 
-    public Map<String, Integer> items() {
+    public Map<String, BigInteger> items() {
         return items;
     }
 
-    public Map<String, Integer> tags() {
+    public Map<String, BigInteger> tags() {
         return tags;
     }
 
-    private static Map<String, Integer> immutableCopy(Map<String, Integer> values) {
+    private static Map<String, BigInteger> immutableCopy(Map<String, BigInteger> values) {
         return Collections.unmodifiableMap(new LinkedHashMap<>(values));
     }
 }
